v0.1.1

+++ Rakefile
@@ -39,13 +39,22 @@ end
 #############################################################################
 
 task :release do
+  puts ""
+  puts "Are you sure you want to relase Timeshifter #{Timeshifter::VERSION}?"
+  print "[y/N] "
+  exit unless STDIN.gets.index(/y/i) == 0
+  
   unless `git branch` =~ /^\* master$/
     puts "You must be on the master branch to release!"
     exit!
   end
+  
+  # Build gem and upload
   sh "gem build timeshifter.gemspec"
-  #sh "gem push timeshifter-#{Timeshifter::VERSION}.gem"
+  sh "gem push timeshifter-#{Timeshifter::VERSION}.gem"
   sh "rm timeshifter-#{Timeshifter::VERSION}.gem"
+  
+  # Commit
   sh "git commit --allow-empty -a -m 'v#{Timeshifter::VERSION}'"
   sh "git tag v#{Timeshifter::VERSION}"
   sh "git push origin master"
